Basic 4.4 class definition compatibility
Map Preview UI is broken in this build!!! You should still be able to get good exports.

This version also includes some unresolved console spam and a more simple approach for indexing multi-bakers, which may come with instability when adding/removing.

Multi-baker support in the Pack Maps panel/UI is visible in this branch, but is still not supported yet.